requirements.txt
yt-dlp
Pillow
qrcode[pil]>=7.4.2
pyzbar>=0.1.9
